convert string id to int

+++ docroot/WEB-INF/src/edu/osu/cws/evals/portlet/TestsAction.java
@@ -36,7 +36,7 @@ public class TestsAction implements ActionInterface {
     public String deleteAppraisal(PortletRequest request, PortletResponse response) throws Exception {
       System.out.println("delete appraisal");
       System.out.println(request.getParameter("id"));
-      Appraisal appraisal = AppraisalMgr.getAppraisal(request.getParameter("id"));
+      Appraisal appraisal = AppraisalMgr.getAppraisal((int)request.getParameter("id"));
       AppraisalMgr.resetAppraisal(appraisal);
 
       return homeAction.display(request, response);
